src/requirements.txt
jsonschema==4.21.1
colorlog==6.8.2
sanic==23.12.1
tortoise-orm==0.20.0
faker==23.3.0
APScheduler==3.10.4
aiomysql==0.2.0
python-dotenv==1.0.1
pyfiglet==1.0.2